refactor: Wish Repo에 SRP적용
