+++ src/main/java/com/example/demo/repository/wish/WishRepositoryImpl.java
@@ -25,25 +25,6 @@ public class WishRepositoryImpl implements WishRepository{
               .update();
   }
 
-  @Override
-  public boolean existProductById(Long productId) {
-    String sql = "SELECT COUNT(*) FROM product WHERE id = :productId";
-    Long count = jdbcClient.sql(sql)
-                           .param("productId", productId)
-                           .query(Long.class)
-                           .single();
-    return count > 0;
-  }
-
-  @Override
-  public boolean existUserById(Long userId) {
-    String sql = "SELECT COUNT(*) FROM users WHERE id = :userId";
-    Long count = jdbcClient.sql(sql)
-                           .param("userId", userId)
-                           .query(Long.class)
-                           .single();
-    return count > 0;
-  }
 
   public void deleteWishProduct(Long userId, Long productId){
     String sql = "DELETE FROM wish WHERE user_id = :userId AND product_id = :productId;";
